backend/requirements.txt
# Backend dependencies (FastAPI)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Keep existing Supabase integration
supabase==2.3.5
gotrue==2.4.2
gunicorn==21.2.0
python-multipart>=0.0.6
